requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1
python-multipart==0.0.18
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
